Extend the merge rollup task capabilities
Adds the capability to erase dimension values from a merged segment
before rollup to reduce cardinality and increase the degree to which
common dimension coordinates are aggregated.  This can result in a
space saving for some dimensions which are not important in historic
data.

See: https://github.com/apache/pinot/issues/14310